create migrations add columns restritivo and posicional more alter column...

See merge request neto/backend-node-adonis!9

+++ app/Controllers/Http/ProfileController.js
@@ -2,6 +2,7 @@
 
 const Profile = use('App/Models/Profile')
 const User = use('App/Models/User')
+const Message = require('../../Hooks/Message')
 
 /** @typedef {import('@adonisjs/framework/src/Request')} Request */
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
@@ -32,8 +33,8 @@ class ProfileController {
    * Create/save a new profile.
    */
 
-    const dados = request.only(['profile', 'unidade', 'carteira', 'rules'])
-    
+    const dados = request.only(['profile', 'unidade', 'carteira', 'rules', 'restritivo', 'posicional'])
+
     const user = await User.find(params.user_id)
 
     const profile = await Profile.findOrCreate(dados)
@@ -43,6 +44,7 @@ class ProfileController {
     profile.user = await profile.users().fetch()
 
     return user
+    //return profile
 
   }
 
@@ -71,9 +73,9 @@ class ProfileController {
   async update({ params, request }) {
 
     const data = request.only([
-      'profile'
+      'profile', 'unidade', 'carteira', 'rules', 'restritivo', 'posicional'
     ])
-
+    console.log(data)
     const profile = await Profile.find(params.id)
 
     profile.merge(data)
@@ -90,9 +92,17 @@ class ProfileController {
   async destroy({ params, response }) {
 
     const profile = await Profile.find(params.id)
-    await profile.delete()
 
-    return Message.messageOk('deleted')
+    if (profile) {
+      await profile.delete()
+      return Message.messageOk('deleted')
+
+    } else {
+      return Message.messageNotFound(`Not Found profile ${params.id}`)
+    }
+
+
+
   }
 }
 module.exports = ProfileController
